requirements.txt
luma.core==2.4.2
luma.oled==3.13.0
numpy==2.1.0
Pillow==10.4.0
PyAudio==0.2.14
setuptools==66.1.1
stupidArtnet==1.5.0
stupidArtnet==1.5.0
